split portfolio in multiple files
