+++ tests/integration/services/portfolio/test_portfolio.py
@@ -4,7 +4,8 @@ from alphaswarm.config import ChainConfig, Config, WalletInfo
 from alphaswarm.services.alchemy import AlchemyClient
 from alphaswarm.services.chains import EVMClient
 from alphaswarm.services.portfolio import Portfolio
-from alphaswarm.services.portfolio.portfolio import PortfolioBase, PortfolioEvm
+from alphaswarm.services.portfolio.portfolio_evm import PortfolioEvm
+from alphaswarm.services.portfolio.portfolio_base import PortfolioBase
 
 
 @pytest.fixture
